move ConfirmationModal into its own file
