+++ src/holofuel/pages/Inbox/ConfirmationModal/index.js
@@ -1,0 +1,3 @@
+import Component from './ConfirmationModal'
+
+export default Component
